Fixed Problems 61/63+65
